Added verification dialogs
Added verification dialogs before sending door and light commands. This addresses the issue of the
app coming up with a button focused and accidentally sending a command. Such an accident
caused the coop light to be on one evening when it should not have been. The chickens were confused
by darkness outside but light inside.